xmonad: Fix tracking last focused tab when floating windows present
This also simplifies the Limit modifier by removing the focus tracking
code, and improves behavior in the rare case when an unfocused tab is
closed. However, the change to tracking by window id instead of index
means the behavior when adding a new tab is changed. When no tab has
been focused before, it still focuses the new tab, but otherwise it
maintains the focused tab.