nvim: Fix code lens with newer versions of Haskell language server
The commands are no longer included with the initial response, and must
be resolved asynchronously. I reimplemented the resolution instead of
just overriding lsp.codelens.display. The latter would have resulted in
flickering, because the lenses are first all displayed as "Unresolved
lens ...", then individually when resolved, and then redisplayed when
all are resolved. I could control the first and third from
lsp.codelens.display, but not the second.