nvim: Sandbox Haskell language server
I have tested that it works with Cabal projects (which require network
and read-write access to ~/.cabal) in a GHCup installation (requiring
read access to ~/.ghcup). Even without sandboxing, I have not had
success with Stack projects, but in theory it should require network and
read-write access to ~/.stack, as well as read-write access to ~/.ghcup
if Stack uses GHCup to download GHC.

Opening a project takes very long each time unless the caches (hie-bios
in particular) are persisted. I don't persist it for single-file
projects (no root_dir) though, since that is fast enough and I don't
want to clutter the cache.