nvim: Only apply LSP document highglights in normal mode
With the default colorscheme and mine since the parent commit, it looks
the same as the visual selection, and it isn't really useful in other
modes anyway. I don't clear it in command mode though, since it visually
only clears when leaving command mode, and then just highlights it
again.