nvim: Fix statusline path for narrow windows and dirvish
The previous implementation would truncate the path in narrow windows,
even when used for titlestring or tabline. It also didn't use a relative
path for dirvish buffers (probably due to the trailing slash).

Another effect of this change is that help buffers now show the full
path. I could add an o.buftype == "help" check for the first branch, but
that would be less consistent.