highlight: Remove configuration
This was mainly about generating and using a theme based on the nvim
colorscheme. It's currently broken, but would be easy to fix. However,
currently nvim is using a separate palette from the main terminal
colors, so I need to reconsider whether I want this consistency and how
to achieve it (and whether I want to use bat instead for syntax
highlighting in ranger).