nvim: Add snippet mappings and change tabstop highlight
I tell language servers that snippets are not supported, since I don't
want functions to be completed as snippets with the parameters as
tabstops (jumping between tabstops is no less work than typing
parentheses and commas). I don't get the snippets if I don't accept the
completion with <C-y> (which this commit also changes to accept the
first completion when none is selected), but sometimes that is necessary
to auto-import a module.

However, some language servers still send other snippets (like the Lua
server's snippets for type annotation comments, and postfix snippets on
@), which may even be useful, so I'm adding mappings to use them.