nvim: Remove terminal mappings and auto terminal mode on focus
It's easy enough to open a terminal using the :te command, which can
also be prefixed by vert or hor for a split. Not automatically entering
terminal mode is more consistent with other buffers, and prevents having
to define window-switching mappings in terminal mode (which I have also
removed here).