nvim: Don't automatically show foldcolumn
With some foldmethods (typically expr), there can be deep nesting (e.g.
in the :InspectTree window). It then uses 9 columns, and even that may
not be enough, so that it starts showing digits in the foldcolumn. In
addition to being noisy, there is no separation between these digits and
the line number. That can be resolved by changing the highlight of one
or the other, or by increasing numberwidth, but I don't want to do that
globally just for the few cases with deeply-nested folds.

I think foldcolumn should rather be set on a case-by-case basis wherever
foldmethod is set. I'm not setting it for my config files that use fold
markers, since the location of the folds can be determined without the
foldcolumn (which is also true for most foldmethods). I'm leaving it for
diffs, because that otherwise lacks a clear indication of where the
folds are.