nvim: Add more type annotations
I'm trying to annotate most parameters and other variables that the
language server fails to infer a type for.

This also includes a fix for an annotation in haskell.lua. It turns out
that with type table<a, b>, the language server doesn't realize that
indexing it may produce nil, so I changed it to table<a, b?>.