nvim: Improve type annotations for nvim_create_autocmd and update calls
The provided type annotations use type any for some fields of the
options table - notably the callback, so type inference doesn't work for
its parameter. This commit introduces more precise types.

I have renamed the callback's parameter to args. I got the name ev from
the help for nvim_create_autocmd, but I prefer args and found that it is
more common in the nvim runtime files.

I have also updated some calls to pass patterns as tables and commands
as strings, now that I know it's an option.