nvim: Don't override locally-changed options when sourcing init.lua
Previously, when manually sourcing init.vim, it would enforce my
preferred settings for e.g. indentation, showing whitespace and line
numbers, even when the current buffer/window explicitly uses different
settings. Now it only sets local option values if they are unchanged
from the global value, in addition to always setting the global value.

It only sets the local value for the current buffer/window, so in the
event that an option is changed in init.lua, all other buffers/windows
will now be considered changed from the default, and will not be updated
even when sourced from that buffer/window. However, it should cause
problems much less frequently than the issue described above, which
happens regardless of whether any options were changed in init.lua.

An alternative solution would be to move this configuration to a plugin
script or Lua module, so that it is only reloaded (manually) when
necessary.

I have also removed the :Source command (to source a script in a new
buffer, window and tab), which was previously used to solve the same
problem.