bin: Improve venv scripts
As an alternative to giving the path to the venv, it is now possible to
give the path to a project (default .) that contains a venv named .venv
(the same name used in the LSP setup).

The functionality of the venv-jupyter-lab script has been moved to the
venv and jupyter-lab scripts, and is now accessible by running "venv
<path> jupyter-lab". It still runs in a sandbox (so that the ipykernel
is not installed globally), but it now binds both the project directory
and venv writably, to allow editing notebooks and installing packages.
The sandboxing support seems like it could be useful in other contexts,
so it is now accessible using "venv <path> sandbox <args>".