bin: Fix lock script
The "systemd --user" service manager now has its own logind session with
class "manager" (as opposed to "user" for regular sessions). loginctl
produces an error when trying to lock the manager session, and if it is
before any user sessions, those are not locked either. Based on the
logind source, the only classes that can be locked are SESSION_USER and
SESSION_USER_EARLY, but I don't know how the latter is represented in
loginctl's output, and it sounds like it's intended for root sessions,
so my solutions is to only attempt to lock "user" sessions.

The "loginctl lock-sessions" (as opposed to lock-session) command
doesn't have this problem, but it doesn't work as a substitute because
it requires authentication (presumably because it may also lock other
users' sessions).

The default tabular output of loginctl doesn't seem intended to be
machine-parsable, and a regex can easily break if the columns are
changed or a cell contains a space, so I have switched to the JSON
output, although this introduces a dependency on jq.