nvim: Better distinguish empty and nonexistent lines in diff
It was possible to distinguish before, by looking for the presence of a
line number, or whether the entire line is highlighted in the the other
buffer, but this makes it easier to distinguish. Nonexistent lines
(DiffDelete) are now a darker black than the background, as if they were
cut out of the buffer. I avoided red, as it isn't necessary to draw
attention.