picom: Switch to window rules and simplify config
Apparently window rules are now recommended over shadow-exclude,
clip-shadow-above and opacity-rule.

I removed the shadow radius and offset, because it was not significantly
different from the default values.

I also removed unredir-if-possible, because it almost never applies
(only when the fullscreen window is covering the bar and there is only
one monitor), and isn't always desired (it disables vsync).