picom: Fix rule updates when windows toggle between tiled and floating
The @ tells picom to look up the _XMONAD_TAGS property on the client
window instead of the frame window. xmonad doesn't reparent windows, so
this shouldn't make a difference, and in fact it succeeds in reading the
property when the window is first created. But it keeps track of which
properties have changed to only reevaluate the window rules when
necessary, and while it used to mark it as changed on the frame window,
since commit 518f63b9200759a46bc8f64b891ffb52451e5275 it is marked as
changed on the client.