nvim, zathura: Fix synctex forward search
g:vimtex_view_zathura_options is only used when initially launching the
viewer, not when performing forward search, so the search was running in
the sandbox. Change it to use the presence of either zathura's forward
or inverse search options (the latter is passed on initial launch, and
the former when searching) to decide whether to use the sandbox.