latexmk: Update clean patterns for vimtex temp output files
vimtex now uses a different naming scheme.

latexmk v4.85 (not yet included in Arch's texlive packages) introduces
$out2_dir, which can be used to keep the intermediate outputs in the
build directory and only copy the final outputs to the working
directory. It's a cleaner solution to the same problem (the viewer
loading intermediate outputs) than g:vimtex_view_use_temp_files, since
it doesn't create extra files in the working directory. However, vimtex
will also need to be updated to understand it, because otherwise it will
just open the output in $out_dir (which would be set to the build
directory).