sandbox: Add support for mounting paths as overlayfs
The -r and -w options are extended so that the source can be a sequence
of colon-separated paths defining the layers from bottom to top. When
the destination path is not specified, it defaults to the path of the
base layer.

For -w, only the top layer is written to. Its path can be empty (if the
layer sequence is empty or ends in a colon), in which case it is a
tmpfs, otherwise writes go to a subdirectory "upper" of the given path,
which is automatically created. There is also a "work" directory, which
must be on the same filesystem, which is why I don't create a temporary
work directory and use the provided path directly as the upper
directory.

Binding device files has also been absorbed into the -w option. It uses
--dev-bind when the source path is in /dev (rather than checking whether
it refers to a device file, to allow binding a directory of device
files). I'm not sure whether this really provides better security than
using --dev-bind unconditionally in place of --bind, since I don't have
device files scattered elsewhere on my system, although at least as it's
implemented it doesn't allow a symlink to a device file to be bound as
the device file.