sandbox: Change -r and -w argument to dest=src instead of src=dest
I more often want to specify the destination path relative to ~ than the
source path, and bash (but not zsh) limits expansion of ~ after = to
when the LHS is an identifier. When mounting a tmpfs, then argument now
ends in =, which has no special meaning in zsh, while an argument
starting with = has to be escaped.

It's also more intuitive as an assignment, and better reflects the fact
that the destination path defaults to the leftmost source path, and with
-w it's the destination (and last source) that's writable, not the first
source.

It's still split so that the destination path can contain = and :, while
the source paths cannot.