kmonad: Release shift when rolling from shifted key
This helps, for example, when typing a letter after _, but before the
underscore key has been released, which would previously incorrectly
capitalize the letter.

It is also possible to use around-when-alone, which presses shift again
if the shifted key is still held when all other keys are released.
However, I don't see much advantage to it, and I have observed it to
cause kmonad to hang.

This is a new feature in kmonad, which I was able to update since PR
524, which I rely on to ensure that tap-hold keys can be reliably used
with modifiers (it delays the release of previous keys until the
tap-hold decision has been made) has been rebased as PR 914. I also
apply the patch from PR 913, which fixes key repeat in the tty.