sandbox,nvim,bin,zsh: Allow specifying literal path to bind in sandbox
The characters = and : are treated specially in the argument to -r or
-w, so that it was impossible to specify a (source) path containing
either character. Now the argument can be prefixed with : to treat the
rest of the argument as a literal path. It is not possible to specify a
different destination path or overlays using this syntax, which I think
is an acceptable limitation given the intended use case.

Either = or : would work best for the prefix, as it assigns meaning to
an argument that would previously have been invalid, while any other
character would prevent binding a file starting with that character
(although it's easy enough to work around by prefixing with ./ or using
the literal syntax). : was chosen over =, because the latter has special
meaning at the start of a word in zsh, although that's a minor concern,
since this feature is mainly intended to be used in scripts binding
unknown paths, where the argument should already be quoted.

This also updates the zsh completion function, and existing uses in nvim
and scripts. In addition to ensuring the scripts work correctly given
arbitrary paths, this also fixes a potential security issue: if a repo
has an LSP root marker at a nested path containing = or :, then opening
a file under that path in nvim could mount an arbitrary path outside the
repo in the LSP server sandbox. A similar thing is possible by making
.venv a symlink, but even with this fix it could be a symlink outside
the repo, which is intended since I want to be able to store the venv
elsewhere.