nvim: Mount site directory in lua LSP sandbox, and allow symlinks
User configuration can go in ~/.config/nvim or ~/.local/share/nvim/site.
I prefer local configuration to go in the latter location, to keep it
separate from the configuration in this repo. I've removed the sourcing
of local_init.vim in favor of this.

With this change, the site directory (and other directories in
'runtimepath', but not their contents) can be a symlink. I've removed
the call to vim.fs.normalize - not only is it unnecessary to normalize
the output of realpath; I don't think it was ever necessary, since
nvim_get_runtime_file returns absolute paths.

All runtime directories are now explicitly mounted in the
lua-language-server sandbox (although it's only the site directory that
was previously unavailable). This isn't strictly necessary, as the
language server works for open buffers that don't correspond to an
accessible file. However, this allows it to search in unopened files as
well.