nvim: Add nvim_ api functions to global scope
Of all the fields of vim, vim.api is the one I use most often (38 uses
so far in init.lua and 12 elsewhere). And since the functions are all
prefixed with nvim_, it's verbose even when aliasing api = nvim.api. For
the same reason, we can add them to the global scope without really
polluting the namespace.

The previous attempt at this was reverted in e48f773, since the
lua-language-server didn't understand it. Now we create a temporary meta
file and pass it to the language server, so it knows that these
functions are available in the global scope. Another difference is that
in the previous version the functions were not available directly in the
global scope, but as fields (with the nvim_ prefix removed) of a global
"nvim" table. The advantage of just placing them directly in the global
scope is that you can more reliably get the correct help tag using K (in
visual mode).

Unlike the previous attempt, the other fields of the vim table are not
added to the global scope. It has many fields, some of which have short
or common names, and only a limited number are used often. The only ones
that are used nearly as often as vim.api are vim.fn, vim.lsp and
vim.keymap.set. These have local aliases in init.vim (with respectively
23, 27 and 34 uses) and fewer uses elsewhere (6, 10 and 3, although 7 of
the vim.lsp uses are in the haskell ftplugin, which has its own local
alias), so I think it's good enough as is, although adding global
aliases just for these fields would also allow for more convenient
completion in command mode. I have removed the local aliases for vim.o
(for consistency with vim.go, vim.bo and vim.wo, and since there are
only 8 uses after introducing the defaults table in 8d9411b) and vim.cmd
(which has even fewer uses at 7).

If I were to add all fields of vim to the global scope in addition to
the api functions, then vim should be set as __index on the metatable,
and the functions assigned in a loop. This is because using __index for
both would require it to be a function rather than a table, which is not
supported in command-line completion (although LSP completion would work
by adding to the meta file), and the vim fields can't be assigned in a
loop, because some are lazily loaded.