nvim: Improve display of quickfix buffer
This adds additional highlights:
- The file path is dimmed in all but the first (consecutive) entry for
  the same file, to make it easier to see how entries are grouped by
  file.
- All diagnostic types are highlighted - not just errors (which are
  changed to the DiagnosticError group, both for consistency and because
  the default qfError highlight is overridden by my QuickFixLine
  highlight).
- The range returned from the language server is highlighted for
  references or multiple declarations/definitions/type
  definitions/implementations.

Line wrapping is disabled, e.g. for long lines matched by grep.

Line numbers are padded to a minimum of 4 characters. By further
omitting the end line number, column numbers and error number, the text
for all entries from the same file will be aligned.

Further differences from the default format are that the default:
- doesn't use ~ to shorten paths outside the current directory, and
  omits the directory for help files returned from :helpgrep (with
  item.type == "\1"),
- replaces the path with item.module if non-empty,
- shows item.pattern instead of item.lnum if the latter is zero,
- displays unknown values of item.type literally, considers "" to be the
  same as "e" if item.nr is nonzero, and
- doesn't trim leading whitespace from item.text if nothing else is
  shown before or between the | characters.
Then first and third points are intentional, but the others can be
aligned with the defaults if I notice anything that doesn't work nicely.