nvim: Remove asyncdo plugin
Since the next thing you want to do is usually to look at the results, I
want to reevaluate whether it's really a problem for :make and :grep to
block the editor.

I'm keeping the :Grep command to run :grep with the silent modifier, but
have changed the <Leader>m mappings to run :make directly, just for
feedback that it's running.