nvim: Improve maps and autocmds for managing quickfix windows
<CR> in a quickfix window now closes it after jumping. <M-CR> can be
used for the original behavior (and can be followed by <M-]> using the
same modifier).

The QuickFixCmdPost autocommand no longer switches focus back to the
previous window. This is unnecessary when the command will jump to the
first entry (e.g. :make or :grep). In other cases (e.g. :make! or
:grep!), one probably wants to look through the results (perhaps
choosing one to jump to with the new <CR> mapping, or filtering with an
fzf mapping I plan to add), so it's better to focus the quickfix window.
This also aligns its behavior with gO and grr.

By a similar argument, the toggle mapping also leaves the focus in the
quickfix window when opening it.

The autocommand and toggle mapping are now also defined for location
list windows.