nvim: Update tag stack from quickfix
Jumping from the quickfix list with <CR> or <M-CR>, but not with the
unimpaired mappings or their underlying commands, will add an entry to
the tag stack.

Goto definition uses the tag stack, as do goto declaration/type
definition/implementation if there is only one match. But if there are
multiple matches, the latter adds them to a quickfix list (and find
references always uses the quickfix list). Then it feels more consistent
if selecting one of the matches from the quickfix list also uses the tag
stack.

In addition, selecting an entry from the quickfix list then visiting
other entries using the unimpaired mappings, can leave many entries on
the jump list. This makes it hard to get back to where you were before
with <C-o>, so it's easier to use the tag stack.

I have also removed the mapping of <C-RightMouse> to <C-o>. By default,
<C-RightMouse> is the same as <C-t>, which seems better as the reverse
of <C-LeftMouse>/<C-]>.