fzf, zsh: Adjust fzf options
Ignoring ctrl-z is now configured in FZF_DEFAULT_OPTS, so that it
applies everywhere. It doesn't look like it's necessary in most places
anymore (in nvim ctrl-z does nothing; in zsh and the unicode-input
script it seems to be automatically resumed after being suspended, since
it redraws the screen and continues working), but in fzf in ranger
pressing ctrl-z suspends ranger, and resuming leaves it in a broken
state.

I've also configured it so that when entries are separated by nulls,
newlines are explicitly shown instead of showing the entry over multiple
lines.