nvim: Reimplement fzf mappings
This implements the files (<Leader>ff), buffers (<Leader>fb) and grep
(<Leader>fg) mappings removed in the parent commit (this time without
relying on the fzf plugin), but not their buffer-relative versions
(<Leader>fF and <Leader>fG). I have removed the rarely-used search
through lines of the buffer (<Leader>f/), and plan to reimplement the
quickfix mapping (<Leader>fq) as a single-key mapping local to quickfix
buffers.

Aside from some minor visual changes, the main functional change is that
there are no longer bindings to open a file in a split (which I rarely
used, and can be achieved by opening a split first), or to create a
quickfix list from the selected lines (which is now automatically done
when more than one line is selected). There are also functional
differences in the buffers mapping (it shows unnamed buffers and
modified state, the delete key has changed and can close windows).