nvim: Update grep configuration
Add rg's -H option, to ensure the filename is printed even when only
searching through one file.

Add a bang to the expansion of :Grep, to match the behavior of other
actions that produce a quickfix list of locations (not :make) that the
quickfix list is focused.

The "grep for word under cursor" mapping is changed to grg, using the
same prefix as LSP mappings, since is it like a version of grr that
doesn't rely on a language server. The buffer-relative version is also
removed, as with the fzf mappings.