nvim: Use rg json output to highlight matches in quickfix
This is cleaner than the previous approach of using escape sequences in
the quickfix text, which is sometimes visible (e.g. when jumping to an
entry with the quickfix window closed) and could theoretically also
affect other quickfix lists not originating from grep.

Unfortunately, this is not as simple as setting grepprg and grepformat,
so :grepadd, :lgrep and :lgrepadd don't automatically work. I would have
to add variants of the command changing the setqflist and copen in the
callback.

The arguments are split by nvim instead of the shell. Notably, spaces
and special characters must be escaped using \ rather than quoted. I
could consider using the shell instead, which would be more consistent
with :IGrep, although it would require escaping symbols with meaning to
the shell (and when expanding %, some characters, like spaces, are
escaped, but others are not).

I have not yet removed the highlighting using escape sequences in
quickfix buffers, because :IGrep still relies on it. However, I plan to
change that to use the json output as well.