nvim: Add workspace symbols mapping using fzf
This could be seen as an fzf mapping (<Leader>fs) or as an LSP mapping
(grs). I chose the former, since other LSP mappings currently default to
the quickfix list, not fzf (see 527997b).

Only the first word of the query is passed through to the language
server (as I had it configured with coc.nvim), while the rest are only
used to filter in fzf. This gives a consistent interface (e.g. improving
on lua-language-server which returns irrelevant results and doesn't sort
them), and allows filtering on the kind, container and file. For servers
that correctly interpret an empty query as requesting all symbols (all
currently configured ones except ccls and pyright), the query can be
started with a space to only apply fzf's filtering. Most servers (except
typescript-language-server) anyway don't support multiple words in the
query. And they either use their own fuzzy search or a stricter prefix
match (haskell-language-server and typescript-language-server), so fzf
shouldn't filter out valid matches.

I have now written my own implementation of both uses of
vim.lsp.util.symbols_to_items, so I could use a local function instead
of overriding that one, but have left it as is for now.