nvim: Replace null bytes in grep output
Quickfix items can't contain null bytes. There shouldn't be any by
default, but could be when using the -a/--text option.

Neither lua patterns (string.gsub) nor vim.regex support nulls in the
pattern (it's interpreted as the end of the string), but vim.re/vim.lpeg
works (as does string.find when the pattern is a plain string, even if
the fourth argument isn't supplied).