nvim: Don't trim leading space in TOC location lists
This is instead of the previous approach of using non-breaking spaces
for indentation to prevent trimming (although the table of contents for
help pages still do this, because I don't control that code). The
advantage is that it also works for man pages, where regular spaces are
used for indentation.