nvim: Change statusline for terminal windows
Show "[Terminal]" followed by the terminal title, instead of the
"term://" buffer name. The title is not shown in the tabline or
titlestring, because those are not automatically updated when the title
is changed. For consistency, hide the quickfix title in those cases as
well.

I would like to combine the absolute and title arguments into an options
table, but the statusline format doesn't allow } in expressions.