nvim, zsh, bin: Rewrite igrep-format from C++ to Lua
It now also uses rg's JSON output (following 9caccca). This change
allows sharing the code for parsing it and formatting for fzf with nvim,
and using nvim's more accurate calculation of string display width.
Another effect is that the matches are now highlighted in green, like
fzf's regular match highlight.

A simple Lua script that implements the same logic as the C++ program
runs about equally fast, using LuaJIT. (These results are based on a
search for a short string through my home directory, in which case the
actual rg process takes two orders of magnitude longer.) The version in
this commit, which parses the JSON and uses the full-featured fzf
formatting (slightly optimized, mostly by avoiding string.gsub), is 7-8
times slower. A potential further optimization, which can get the
overhead down to 5-6 times, is to use the byte length as a first
approximation in the padding calculation, and only call strwidth if that
indicates there is a possibility of a positive amount of padding (it
seems most of the overhead comes from very long strings, where no
padding is necessary).