nvim: Don't move filename to left in fzf if line number is present
It is moved to the left for cases where the filename is the only
information in the quickfix entry, e.g. when selecting from a quickfix
list populated using <Leader>ff or <Leader>fb. However, it would also do
so for e.g. grep results matching an empty line, while it would not be
moved in other (non-empty) lines, which has now been fixed.