nvim: Use fzf to implement vim.ui.select
I have observed rust-analyzer to suggest over 100 code actions, which
requires something like fzf to efficiently select one. For shorter
lists though, the default implementation that allows selecting an item
without filtering or moving the selection, is probably more convenient.
With the previous commit, for lists of fewer than 10 items, it can be
done in the same number of key presses (ctrl-s and a letter, compared to
a digit (on a layer) and enter).