nvim: Allow editing rg arguments from <Leader>fg mapping
Ctrl-o is bound to toggle between editing the pattern and editing
additional arguments. The alternative would be to just interpret the
query string as arguments including the pattern, but that would require
the user to escape special characters in the pattern.

It's no longer necessary to provide the :IGrep command in addition to
the <Leader>fg mapping, since the former was only there to allow
providing additional arguments.