nvim: Prefer vim.startswith/endswith and string.find over string.match
From my previous experiments for igrep_format.lua, vim.startswith and
vim.endswith are always faster, and string.find is faster when the
pattern is a plain string. Performance is not an issue for any of these
instances, and the pattern for string.find is not plain (I've left some
instances of string.match where that feels like a more appropriate
verb), but changing what I use by default can prevent future problems.