nvim: Use shell to split :Grep arguments
I'm more familiar with the shell syntax, this allows using globs, and
it's more consistent with the default :grep and :make (and the :Make I'm
planning to add). However, it may require quoting or escaping the
pattern in more cases.

This also adds a grG mapping to search for a WORD, like grg searches for
a word.