nvim: Incrementally update quickfix list while Make/Grep is running
This allows to see the progress of a build by opening the quickfix
window. It's not opened by default, because that would unnecessarily
open and close it for error-free builds, and opening it gives an obvious
indication that the process is finished.

This also changes :Grep to report errors in the quickfix list instead of
discarding the results, which is useful when it can't open some files
but finds matches in others.