nvim: Allow running an arbitrary command without setting makeprg
This uses the prefix :Make!, so that :Make! can be thought of as a
version of :! that creates a quickfix list from the output.

This uses the global errorformat, since the local one could have been
set for a specific makeprg. There is no need for a variant that doesn't
use errorformat, since in that case it makes more sense to use a
terminal buffer (prefix the command with :hor te for similar behavior).