nvim: Improve behavior of trees (LSP symbols) in quickfix
When tree_mode is enabled for a quickfix list, folded lines are
displayed normally, instead of getting the Folded highlight (which also
fixes horizontal scrolling), and the foldcolumn is enabled to still be
able to identify folds. In addition, the left and right arrows are
mapped to close and open folds.

It is intentional that the tree_mode context is not preserved when
filtering the list using fzf, because after doing so the items may be
reordered and/or missing their parent items.