nvim: Improve behavior when selecting quickfix item using fzf
The previous fix to ensure the help or man window is used when selecting
an entry from its table of contents, wasn't working when moving the
location list window to a sidebar.

As an additional benefit of the new approach, selecting a single entry
also marks it as the selected entry in the quickfix list.