nvim: Update Lua LSP settings
Diagnostics are no longer disabled for unopened files in the workspace,
which is useful when using vim.diagnostic.setqflist. We don't want that
to include diagnostics from plugins though, so exclude those from the
workspace.

The unused-function and empty-block diagnostics are disabled, as they
cause entire blocks of work-in-progress code to be highlighted as
comments. Unused functions are still detected by unused-local, which
only highlights the name.

The global-element detects all unintentional assignments to global
variables (e.g. the one in the Haskell ftplugin), unlike
lowercase-global.