nvim: Implement LSP call and type hierarchy commands
It recursively expands the hierarchy in the given direction (incoming or
outgoing, supertypes or subtypes), and shows the results in the quickfix
list in tree mode.

This also adds filenames to the quickfix list in tree mode. It uses the
faint NonText highlight, because in the more common case of the gO
mappings, it doesn't provide useful information.

For call hierarchies, the server also returns the ranges of the function
calls (as opposed to the function definition). This is ignored, since
there isn't a way to associate multiple ranges with the same quickfix
item, and when using just a call range it would look like the symbol is
defined in the wrong file.

No configured servers support the type hierarchy requests, but it (in
theory) can be supported by the same code as call hierarchies.