nvim: Integrate with REPLs
This adds a g= operator (chosen because nvim master has a g== mapping in
help files with similar functionality), which opens a REPL (configured
in an ftplugin) in a terminal buffer, and sends the selected text to it.

For lua files in the nvim runtime, it is overridden to instead be
executed directly in the editor.