nvim: Preserve locals in lua REPL
Each use of the g= mapping executes its code as a separate chunk, so
locals are by default not preserved across uses. This commit creates an
environment table to preserve top-level locals and implicit globals set
in the REPL. It's still possible to set actual globals using _G (which
is currently how all my code sets globals, if only to indicate it's not
doing so by accident). We could make undeclared variables in the REPL
refer to globals to match the standard behavior, by also setting
__newindex = _G in the metatable and using rawset to add the locals.
However, that would pollute the global scope when evaluating a snippet
of code that excludes a local declaration, or if a previous chunk
declared the local but set it to nil.