nvim: Enable loading file in REPL
This uses the g=s mapping (mnemonic: source or script).

There is a default implementation that just sends all the text in the
buffer to the REPL, but most REPLs provide a command to load a file,
which in some cases (haskell, python and nvim lua) load it in a clean
environment.