bin: Improve altgr and compose sequences in unicode-input
Since the recent change to convert unicode symbols to tex commands in
nvim, I have added the following to ~/.XCompose (and similar for phi):
<Greek_epsilon> : "ϵ"
<Greek_omicron> <Greek_epsilon> : "ε"

This changes <altgr-e> to produce the epsilon symbol, which corresponds
to \epsilon, while I must use <altgr-o><altgr-e> (mnemonic: other
epsilon) for the Greek letter, which produces \varepsilon. There is a
default sequence <compose>ε', so I cannot use <compose>ε, but omicron
works as an alternative, because it appears identical to Latin o and so
isn't used as a variable (there is no \omicron in tex).

This commit changes the unicode-input script to include compose
sequences that don't start with <compose>, showing <compose> explicitly
when present. It also expands altgr in compose sequences (e.g. giving
<altgr-o><altgr-e> instead of οε), and notices when a single-key compose
sequence effectively overrides a key in the keymap (e.g. ε can no longer
be typed with <altgr-e>). Finally, sequences from ~/.XCompose are now
shown before the defaults.