nvim: Automatically choose split direction for REPL window
It uses the same condition as sidebar windows, but isn't a sidebar,
since it can benefit from more width. (Currently there would be value in
fixing its width, since the text is truncated when resized. However, in
nvim 0.11 it will wrap (or scoll in the scrollback) instead, so I'm not
doing that.)