nvim: Remove automatic starting of terminal mode
The main issue solved by this is that entering terminal mode scrolls to
the bottom, so it would scroll whenever focusing a terminal window. It
also allows removing the duplicate window navigation mappings in
terminal mode and the hack to exit insert mode in the REPL code. This
also adds an easier mapping to exit terminal mode.