xmonad: Change default window limit and keys to adjust it
The most common layout change I make is reducing it to one column (and
one window), to get a "tabbed full screen" layout (in preference to the
Fullscreen modifier, which hides the bar and tabs). I reset the layout
to undo this, because increasing to 2 columns only increases the window
limit to 2, not the default 3. However, this process doesn't just change
the window limit, but also loses all information about the number and
width of columns.

By changing the default window limit to 2, I can instead achieve the
tabbed full screen layout by decreasing the window limit once (or more
times if it has been increased), and undo it by increasing the limit
again. This way the column configuration is preserved. The other change
then is to swap the key bindings to adjust the window limit and number
of columns, so that it is the latter (less common now) that requires the
extra ctrl modifier.