nvim: Enable clearing expression in REPL buffer
This automatically sends <C-e><C-u> before a new expression, which
clears the last line, so it only works when the user typed something in
the buffer, not when the last expression sent was incomplete. For this
case we map <Leader>ec to send <C-c> to the REPL, which clears all lines
(or interrupts the running expression). We can't use it automatically
(unless we detect whether there is any text after the prompt), since it
usually prints some indication of the interrupt (and exits the lua
REPL).

For ipython <C-c> cleanly clears the buffer, but we can't use it there
either, because if sent with the first expression it causes the REPL to
exit immediately. Instead we rely on ipython's vi mode. This can then
clear multiple lines, but ipython doesn't have the incomplete expression
problem in the first place, because we force evaluation with <M-CR>
(which is still necessary to avoid waiting for an empty line).