nvim: Fix detection of tables of contents in 0.11
Nvim 0.11 adds new gO mappings for markdown and checkhealth buffers, but
changes the title. Note that for markdown the indentation level is only
increased for every two levels of headings (it's not due to the / 2 in
my code), and there is no longer leading whitespace in the man TOC, so
it lacks a hierachy.